rename os_getreg to os_getWindowsRegistry
